+++ packages/shsp/lib/src/impl/socket/dual/dual_shsp_socket_auto.dart
@@ -1,25 +1,38 @@
 import 'dart:io';
 
 import '../../../../shsp.dart';
+import 'package:singleton_manager/singleton_manager.dart';
 
+@dependencyInjectable
 class DualShspSocketAuto
     extends DualShspSocketMigratable
     implements IDualShspSocketAuto {
-  DualShspSocketAuto(Sockets sockets) : super(sockets);
-
-  DualShspSocketAuto.fromWrappers({
-    IShspSocketWrapper? ipv4Wrapper,
-    IShspSocketWrapper? ipv6Wrapper,
-  }) : super.fromWrappers(
-          ipv4Wrapper: ipv4Wrapper,
-          ipv6Wrapper: ipv6Wrapper,
+  DualShspSocketAuto({
+    @Subkey('ipv4') IShspSocketMigratable? ipv4Migratable,
+    @Subkey('ipv6') IShspSocketMigratable? ipv6Migratable,
+  }) : super(
+          ipv4Migratable: ipv4Migratable,
+          ipv6Migratable: ipv6Migratable,
         );
+
+  // ignore: avoid_unused_constructor_parameters, // GENERATED CODE - DO NOT MODIFY BY HAND
+  factory DualShspSocketAuto.dependencyInjectionFactory({String key = 'default', String subkey = 'default'}) { // GENERATED CODE - DO NOT MODIFY BY HAND
+    final ipv4Migratable = RegistryManager.instance.getInstanceNullable<IShspSocketMigratable>(key: key, subkey: 'ipv4'); // GENERATED CODE - DO NOT MODIFY BY HAND
+    final ipv6Migratable = RegistryManager.instance.getInstanceNullable<IShspSocketMigratable>(key: key, subkey: 'ipv6'); // GENERATED CODE - DO NOT MODIFY BY HAND
+
+    return DualShspSocketAuto( // GENERATED CODE - DO NOT MODIFY BY HAND
+      ipv4Migratable: ipv4Migratable, // GENERATED CODE - DO NOT MODIFY BY HAND
+      ipv6Migratable: ipv6Migratable, // GENERATED CODE - DO NOT MODIFY BY HAND
+    ); // GENERATED CODE - DO NOT MODIFY BY HAND
+  } // GENERATED CODE - DO NOT MODIFY BY HAND
+
+  DualShspSocketAuto.fromSockets(Sockets sockets) : super.fromSockets(sockets);
 
   DualShspSocketAuto.fromMigratable(
     DualShspSocketMigratable migratable
-  ) : super.fromWrappers(
-          ipv4Wrapper: migratable.ipv4SocketWrapper,
-          ipv6Wrapper: migratable.ipv6SocketWrapper,
+  ) : super(
+          ipv4Migratable: migratable.ipv4SocketMigratable,
+          ipv6Migratable: migratable.ipv6SocketMigratable,
         );
 
   static Future<DualShspSocketAuto> create() async {
@@ -28,7 +41,7 @@ class DualShspSocketAuto
     if (ipv4Socket == null && ipv6Socket == null) {
       throw Exception('Failed to bind both IPv4 and IPv6 sockets');
     }
-    return DualShspSocketAuto(Sockets(ipv4SocketImpl: ipv4Socket, ipv6SocketImpl: ipv6Socket));
+    return DualShspSocketAuto.fromSockets(Sockets(ipv4SocketImpl: ipv4Socket, ipv6SocketImpl: ipv6Socket));
   }
 
   @override