+++ packages/shsp/lib/src/interfaces/dual/i_dual_shsp_socket_migratable.dart
@@ -13,7 +13,7 @@ abstract interface class IDualShspSocketMigratable
 
   /// The socket wrapper for the given address family. Defaults to IPv6;
   /// pass [InternetAddressType.IPv4] for the IPv4 wrapper.
-  IShspSocketWrapper getSocketWrapper([
+  IShspSocketMigratable getSocketMigratable([
     InternetAddressType type = InternetAddressType.IPv6,
   ]);
 }