+++ packages/shsp/lib/src/interfaces/dual/i_dual_shsp_socket_wrapper.dart
@@ -1,9 +1,0 @@
-import '../../../shsp.dart';
-
-export 'i_dual_shsp_socket_auto.dart';
-export 'i_dual_shsp_socket_migratable.dart';
-
-abstract interface class IDualShspSocketWrapper
-    implements IDualShspSocketAuto {
-  set internalSocket(IDualShspSocketAuto socket);
-}